--- .github/common/download-examples.sh
@@ -1,6 +1,0 @@
-#!/bin/bash
-
-cd autotest
-python ci_setup.py
-cd ..
-
