+++ .github/workflows/pymake-gcc.yml
@@ -42,13 +42,14 @@ jobs:
         uses: modflowpy/install-gfortran-action@v1
 
       - name: Download examples for pytest runs
+        working-directory: ./autotest
         run: |
-          .github/common/download-examples.sh
+          pytest -v -n=auto -m="regression" --durations=0 ci_setup.py
 
       - name: Run pytest
         working-directory: ./autotest
         run: |
-          pytest -v --dist=loadfile -n=auto -m="base or regression" --durations=0 --cov=pymake --cov-report=xml
+          pytest -v -n=auto -m="base or regression" --durations=0 --cov=pymake --cov-report=xml
 
       - name: Print coverage report before upload
         working-directory: ./autotest
@@ -100,13 +101,14 @@ jobs:
           g++ --version
 
       - name: Download examples for pytest runs
+        working-directory: ./autotest
         run: |
-          .github/common/download-examples.sh
+          pytest -v -n=auto -m="base" --durations=0 ci_setup.py
 
       - name: Run pytest
         working-directory: ./autotest
         run: |
-          pytest -v --dist=loadfile -n=auto -m="base" --durations=0 --cov=pymake --cov-report=xml
+          pytest -v -n=auto -m="base" --durations=0 --cov=pymake --cov-report=xml
 
       - name: Print coverage report before upload
         working-directory: ./autotest
@@ -152,4 +154,4 @@ jobs:
       - name: Run pytest
         working-directory: ./autotest
         run: |
-          pytest -v --dist=loadfile -n=auto -m="base" --durations=0
+          pytest -v -n=auto -m="base" --durations=0
